Ajout de la classe Book + test d'insertion dans BDD

+++ 14 - PHP_MongoDB/TP/library-manager/public/index.php
@@ -4,7 +4,7 @@ require_once __DIR__ . '/../vendor/autoload.php';
 
 use Library\Manager\Database\MySQLConnection;
 use Library\Manager\Models\User;
-
+use Library\Manager\Models\Book;
 
 
 // Test 1
@@ -13,9 +13,25 @@ $pdo = MySQLConnection::getConnection();
 echo "<br>Connexion MySQL OK !";
 
 
-// Test 2
+// Test connexion MySQL
 
-$newUser = new User("Alice Dupont", "alice@example.com");
-$newUser->save();
+$existing = User::findByEmail("alice@example.com");
 
-echo "<br>Nouvel utilisateur ajouté avec ID : " . $newUser->getId();
+if (!$existing) {
+    $newUser = new User("Alice Dupont", "alice@example.com");
+    $newUser->save();
+    echo "<br>Nouvel utilisateur ajouté avec ID : " . $newUser->getId();
+} else {
+    echo "<br>L'utilisateur existe déjà avec ID : " . $existing->getId();
+}
+
+// Test d'insertion d'un livre
+$nouveauLivre = new Book(
+    "Le Meilleur des Mondes",
+    "Aldous Huxley",
+    "9782070368228",
+    3 // ID utilisateur déjà existant (ex. Alice avec ID 3)
+);
+
+$nouveauLivre->save();
+echo "<br>Livre ajouté avec ID : " . $nouveauLivre->getId();
